chore(blockifier): move state reader's cairo native logic test to blockifier
