+++ crates/blockifier/src/state/state_reader_and_contract_manager.rs
@@ -7,6 +7,10 @@ use crate::state::contract_class_manager::ContractClassManager;
 use crate::state::global_cache::CompiledClasses;
 use crate::state::state_api::{StateReader, StateResult};
 
+#[cfg(test)]
+#[path = "state_reader_and_contract_manager_test.rs"]
+pub mod state_reader_and_contract_manager_test;
+
 pub trait FetchCompiliedClasses: StateReader {
     fn get_compiled_classes(&self, class_hash: ClassHash) -> StateResult<CompiledClasses>;
 }
